requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
opencv-contrib-python==4.*
opencv-python==4.*
numpy==1.26.4
matplotlib==3.*
scikit-learn==1.6.*
tensorflow==2.20.*
keras
split-folders==0.5.1
keras_tuner==1.4.7
optuna==4.*
patchify==0.2.3
torch==2.7.1+cu128
torchvision==0.22.1+cu128
transformers==4.*
datasets==4.*
pillow==11.*
tensorboard==2.*
pandas==2.*
catboost==1.2.8
imbalanced-learn==0.14.*
nibabel==5.3.*
pyglet==2.*
scikit-image==0.25.2
scipy==1.*
seaborn==0.13.2
shutup==0.3.*
tqdm==4.67.1
trimesh==4.11.*
xgboost==3.1.*
medmnist==3.0.2
huggingface-hub==0.36.*
sentence_transformers==5.*
pymupdf==1.26.*
openslide-bin==4.0.0.*
openslide-python==1.4.*
spams-bin==2.6.*
PyWavelets==1.8.0
imagehash==4.3.2
brisque==0.0.17
simpleitk==2.5.*
PyPDF2==3.0.1
shap==0.48.0
lime==0.2.*
shapely==2.1.*
